refactor: clean up and improve code

+++ src/artist/artist.service.ts
@@ -1,28 +1,12 @@
-import {
-  forwardRef,
-  Inject,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
-import { FavsService } from 'src/favs/favs.service';
-import { AlbumService } from 'src/album/album.service';
-import { TrackService } from 'src/track/track.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { parseError } from 'src/utils/errors';
 
 @Injectable()
 export class ArtistService {
-  constructor(
-    @Inject(forwardRef(() => FavsService))
-    private readonly favsService: FavsService,
-    @Inject(forwardRef(() => AlbumService))
-    private readonly albumService: AlbumService,
-    @Inject(forwardRef(() => TrackService))
-    private readonly trackService: TrackService,
-    private readonly prisma: PrismaService,
-  ) {}
+  constructor(private readonly prisma: PrismaService) {}
 
   async create(createArtistDto: CreateArtistDto) {
     try {
@@ -51,7 +35,12 @@ export class ArtistService {
   }
 
   async update(id: string, updateArtistDto: UpdateArtistDto) {
-    await this.findOne(id);
+    const artist = await this.prisma.artist.findUnique({
+      where: { id },
+    });
+    if (!artist) {
+      throw new NotFoundException(`Artist with id "${id}" not found`);
+    }
     try {
       const updatedArtist = await this.prisma.artist.update({
         where: { id },
@@ -64,17 +53,15 @@ export class ArtistService {
   }
 
   async remove(id: string) {
-    await this.findOne(id);
-    try {
-      //this.favsService.removeArtist(id, true);
-      this.albumService.setNullArtist(id);
-      this.trackService.setNullArtist(id);
-      await this.prisma.artist.delete({
-        where: { id },
-      });
-      return true;
-    } catch (error) {
-      throw parseError(error);
+    const artist = await this.prisma.artist.findUnique({
+      where: { id },
+    });
+    if (!artist) {
+      throw new NotFoundException(`Artist with id "${id}" not found`);
     }
+    await this.prisma.artist.delete({
+      where: { id },
+    });
+    return true;
   }
 }
