+++ src/track/track.service.ts
@@ -1,22 +1,12 @@
-import {
-  forwardRef,
-  Inject,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
-import { FavsService } from 'src/favs/favs.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { parseError } from 'src/utils/errors';
 
 @Injectable()
 export class TrackService {
-  constructor(
-    @Inject(forwardRef(() => FavsService))
-    private readonly favsService: FavsService,
-    private readonly prisma: PrismaService,
-  ) {}
+  constructor(private readonly prisma: PrismaService) {}
 
   async create(createTrackDto: CreateTrackDto) {
     try {
@@ -44,22 +34,13 @@ export class TrackService {
     return track;
   }
 
-  async setNullArtist(id: string) {
-    await this.prisma.track.updateMany({
-      where: { artistId: id },
-      data: { artistId: null },
-    });
-  }
-
-  async setNullAlbum(id: string) {
-    await this.prisma.track.updateMany({
-      where: { albumId: id },
-      data: { albumId: null },
-    });
-  }
-
   async update(id: string, updateTrackDto: UpdateTrackDto) {
-    await this.findOne(id);
+    const track = await this.prisma.track.findUnique({
+      where: { id },
+    });
+    if (!track) {
+      throw new NotFoundException(`Track with id "${id}" not found`);
+    }
     try {
       const updatedTrack = await this.prisma.track.update({
         where: { id },
@@ -72,9 +53,13 @@ export class TrackService {
   }
 
   async remove(id: string) {
-    await this.findOne(id);
+    const track = await this.prisma.track.findUnique({
+      where: { id },
+    });
+    if (!track) {
+      throw new NotFoundException(`Track with id "${id}" not found`);
+    }
     try {
-      // this.favsService.removeTrack(id, true);
       await this.prisma.track.delete({
         where: { id },
       });
